polymarket_arbitrage/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
web3==6.11.3
eth-account==0.10.0
aiohttp==3.9.1
numpy==1.24.3
pandas==2.1.3
sqlalchemy==2.0.23
